requirements.txt
Flask==2.2.2
sqlalchemy==1.4.40
flask-sqlalchemy==2.5.1
flask-wtf ==1.0.1
flask-bcrypt==1.0.1
flask-login==0.6.2
email_validator==1.2.1
Pillow==9.2.0
itsdangerous==2.0.1
flask-mail==0.9.1
psycopg2==2.9.3
sqlalchemy-utils==0.38.3
python-dotenv==0.21.0